Keep same file nodes
Instead of deleting the file and moving a new one in its place, use redirection to just overwrite the original contents.  This avoids the need to try to re-set the file permissions;

Also don't try to re-set the last modified time, and don't try to remove the directory.  Not sure why I ever thought that latter part made sense.  We try to delete empty directories when moving the files, earlier in the script; but when just modifying the contents, of course the directory will not be empty.