Let git tell us which files are binary
There are a few changes in this commit, but the main one is, instead of trying to figure out which files are binary (so we can avoid trying to edit them), go ahead and edit them, and then when we do the diff, git will tell us that it thinks they're binary files.  If we get that feedback, abort the edit.